backend/requirements.txt
alembic==1.14.1
fastapi==0.129.0
uvicorn==0.40.0
pydantic-settings==2.7.1
sqlalchemy==2.0.46
psycopg[binary]>=3.2.4
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
python-dotenv==1.2.1
requests==2.32.3
email-validator==2.3.0
supabase>=2.11.0
pyiceberg>=0.8.0
